# src/main/java/com/vass/shop/infrastructure/adapters/inbound/rest/api/ProductsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.23.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.vass.shop.infrastructure.adapters.inbound.rest.api;

import com.vass.shop.infrastructure.adapters.inbound.rest.dto.ErrorModel;
import org.springframework.lang.Nullable;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.ProductRequest;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.ProductResponse;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-08T22:59:31.432094100+03:00[Europe/Bucharest]", comments = "Generator version: 7.23.0")
@Validated
@Tag(name = "products", description = "the products API")
public interface ProductsApi {

    String PATH_CREATE_PRODUCT = "/products";
    /**
     * POST /products : Create a product
     *
     * @param productRequest  (required)
     * @return Product created (status code 201)
     *         or Invalid request payloads or schema errors. (status code 400)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Create a product",
        responses = {
            @ApiResponse(responseCode = "201", description = "Product created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payloads or schema errors.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ProductsApi.PATH_CREATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ProductResponse> createProduct(
        @Parameter(name = "ProductRequest", description = "", required = true) @Valid @RequestBody ProductRequest productRequest
    );


    String PATH_DELETE_PRODUCT = "/products/{productId}";
    /**
     * DELETE /products/{productId} : Delete a product
     *
     * @param productId  (required)
     * @return Product deleted successfully (status code 204)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "deleteProduct",
        summary = "Delete a product",
        responses = {
            @ApiResponse(responseCode = "204", description = "Product deleted successfully"),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ProductsApi.PATH_DELETE_PRODUCT,
        produces = { "application/json" }
    )
    ResponseEntity<Void> deleteProduct(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId
    );


    String PATH_GET_PRODUCT_BY_ID = "/products/{productId}";
    /**
     * GET /products/{productId} : Get product by ID
     *
     * @param productId  (required)
     * @return Product details found (status code 200)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Get product by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product details found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_GET_PRODUCT_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<ProductResponse> getProductById(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId
    );


    String PATH_LIST_PRODUCTS = "/products";
    /**
     * GET /products : List all products
     *
     * @param limit How many products to return at one time (optional, default to 10)
     * @return A list of products (status code 200)
     *         or Invalid request payloads or schema errors. (status code 400)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "listProducts",
        summary = "List all products",
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of products", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductResponse.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payloads or schema errors.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_LIST_PRODUCTS,
        produces = { "application/json" }
    )
    ResponseEntity<List<ProductResponse>> listProducts(
        @Parameter(name = "limit", description = "How many products to return at one time", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "10") Integer limit
    );


    String PATH_UPDATE_PRODUCT = "/products/{productId}";
    /**
     * PUT /products/{productId} : Update a product
     *
     * @param productId  (required)
     * @param productRequest  (required)
     * @return Product updated successfully (status code 200)
     *         or Invalid request payloads or schema errors. (status code 400)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "updateProduct",
        summary = "Update a product",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payloads or schema errors.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ProductsApi.PATH_UPDATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ProductResponse> updateProduct(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId,
        @Parameter(name = "ProductRequest", description = "", required = true) @Valid @RequestBody ProductRequest productRequest
    );

}