# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Shopping Cart API
  version: 1.0.0
  description: Secure API-First Shopping Cart Service adhering to Hexagonal Architecture constraints.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Server
paths:
...
  /products:
    post:
      summary: Create a product
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    get:
      summary: List all products
      operationId: listProducts
      security:
        - BearerAuth: [ ]
      parameters:
        - name: limit
          in: query
          description: How many products to return at one time
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /products/{productId}:
    get:
      summary: Get product by ID
      operationId: getProductById
      security:
        - BearerAuth: [ ]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    put:
      summary: Update a product
      operationId: updateProduct
      security:
        - BearerAuth: [ ]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    delete:
      summary: Delete a product
      operationId: deleteProduct
      security:
        - BearerAuth: [ ]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted successfully
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Secure all routes using a validated JWT Token.
  responses:
    400BadRequest:
      description: Invalid request payloads or schema errors.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    401Unauthorized:
      description: Authorization header missing or validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    404NotFound:
      description: Requested domain entity could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    500InternalServerError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
  schemas:
    ErrorModel:
      type: object
      required: [code, message, timestamp]
      additionalProperties: false
      properties:
        code:
          type: integer
          minimum: 400
          maximum: 599
        message:
          type: string
          maxLength: 255
          pattern: '^[a-zA-Z0-9 ]+$'
        timestamp:
          type: string
          format: date-time
...
    ProductRequest:
      type: object
      required: [name, price]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        price:
          type: number
          minimum: 0.01
          maximum: 100000.00
    ProductResponse:
      type: object
      required: [id, name, price]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
...